refactor(editor): extract section order panel
Mechanical module split: section ordering (drag & drop), visibility
toggles/rules, hidden headings and the weather/energy sub-options move
to panels/SectionOrderPanel.ts. Moved-line hardening per CLAUDE.md
Codacy pitfalls: HeadingKey typing instead of 'as any', Reflect
get/set/deleteProperty for the section_visibility record.
